requirements.txt
Django>=1.5
django-annoying>=0.7
django-extensions>=1.1
django-extra-views>=0.6
django-debug-toolbar>=0.9
django-debug-toolbar-template-timings>=0.4
django-dajaxice>=0.5
django-celery>=3.0
django-social-auth>=0.7
djangorestframework>=2.0
Werkzeug>=0.8
pygments>=1.6
